Fix Gravity
Previous, gravity was applied so long as the smaller object was
aligned on EITHER x or y axis, which meant that gravity continued
to apply long after the smaller massive body was far away,
so long as it was similarly positioned on either one axis.

Now gravity has a square of influence of 3 * mass of larger body
in each direction.